\name{bed.sample}
\docType{data}
\alias{bed.sample}
\title{A sample of 25 rows from a .bed file of mm9 regions}
\description{25 regions from a .bed file for use in example code, contains regions from mm9}
\usage{bed.sample}
\format{A data.frame with 25 obs. of 3 variables: chromosome, start position, end position}
\keyword{datasets}
