\name{gata4}
\docType{data}
\alias{gata4}
\title{A .bed file containing genomic ranges sampled from gata4}
\description{1000 genomic ranges sampled from gata4 binding sites in mm9}
\usage{gata4}
\format{A data frame with 1000 observations of 3 variables: chromosome, start position, end position.}
\source{He, A. (2011) \emph{Co-occupancy by multiple cardiac transcription factors identifies transcriptional enhances active in heart}. PNAS 108(14), 5632-5637}