\name{p300}
\docType{data}
\alias{p300}
\title{A .bed file containing genomic ranges sampled from p300}
\description{1000 genomic ranges sampled from p300 binding sites identified in HL-1 cells (mm9)}
\usage{p300}
\format{A data frame with 1000 observations of 3 variables: chromosome, start position, end position.}
\source{He, A. (2011) \emph{Co-occupancy by multiple cardiac transcription factors identifies transcriptional enhances active in heart}. PNAS 108(14), 5632-5637}