% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPipeline.R
\name{PCAplot}
\alias{PCAplot}
\title{Plot PCA given an input list of fnAnot charts}
\usage{
PCAplot(input)
}
\arguments{
\item{input}{A list of functional annotation charts.}
}
\description{
Given a list of functional annotation charts, 
this function outputs a PCA plot
}

