% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.PCA}
\alias{plot.PCA}
\title{Plot PCA given an input list of fnAnot charts}
\usage{
plot.PCA(input)
}
\arguments{
\item{input}{A list of functional annotation charts.}
}
\description{
Given a list of functional annotation charts, 
this function outputs a PCA plot
}

