% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPipeline.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Plot dendrogram given an input list of fnAnot charts}
\usage{
plotDendrogram(input)
}
\arguments{
\item{input}{A list of functional annotation charts.}
}
\description{
Given a list of functional annotation charts, 
this function outputs a dendrogram
}

