% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotInteractive}
\alias{plotInteractive}
\title{Interactive plotting function for groups of GO terms}
\usage{
plotInteractive(input, outDir = NULL, prefix = NULL, pdf = TRUE)
}
\arguments{
\item{input}{A list of functional annotation charts.}

\item{outDir}{The directory to save plots to.}

\item{prefix}{The prefix to append to each file, if any.}

\item{pdf}{If true, plots will be pdfs. If false, pngs.}
}
\description{
Given a list of functional annotation charts and optionally an
output directory, this function can output dendrograms, PCA analysis plots and a
correlation matrix to make large-scale comparisons easy.
}

