\name{plotPairwise}
\alias{plotPairwise}
\title{Generates a scatterplot of two sets of GO terms based on DAVID P-values}
\usage{
plotPairwise(setA, setB, cutoff = NULL, useRawPvals = FALSE,
  plotNA = TRUE, model = "lm", ontology = NULL)
}
\arguments{
  \item{setA}{DAVIDFunctionalAnnotationChart object to
  compare}

  \item{setB}{DAVIDFunctionalAnnotationChart object to
  compare}

  \item{cutoff}{The p-value or adjusted p-value to use as a
  cutoff}

  \item{useRawPvals}{If false, uses adjusted p-values,
  otherwise uses the raw ones}

  \item{plotNA}{If true, any GO term present in only one
  list is considered to have a p-value of 1 in the other;
  otherwise, it is simply removed}

  \item{model}{The model to use when plotting linear fit,
  default 'lm'}

  \item{ontology}{If a specific ontology (MF, BP, CC) is
  wanted rather than all terms, supply it here as a string}
}
\description{
Generates a -log10 scatterplot of two sets of GO terms by
p-value or corrected p-value with linear fit and
correlation. Also includes a Jaccard metric for gene
overlap within each GO term. Useful as an overall metric of
gene list similarity. NOTE: The plotZScores function is
more statistically sound, you should use that instead of
this.
}
\examples{
data(funChart1)
data(funChart2)
plotPairwise(funChart1, funChart2)
}

