\name{SRF}
\docType{data}
\alias{SRF}
\title{A .bed file containing genomic ranges sampled from SRF}
\description{1000 genomic ranges sampled from SRF binding sites in mm9}
\usage{SRF}
\format{A data frame with 1000 observations of 3 variables: chromosome, start position, end position.}
\source{He, A. (2011) \emph{Co-occupancy by multiple cardiac transcription factors identifies transcriptional enhances active in heart}. PNAS 108(14), 5632-5637}