\name{ucsc.mm9}
\docType{data}
\alias{ucsc.mm9}
\title{A database dump of the UCSC mm9 refGene track for use in annotation}
\description{Contains one entry per gene with information including chromosome, transcript start/end and strand}
\usage{ucsc.mm9}
\format{A data.frame with 32608 obs. of 16 variables, including chromosome, transcription start and end and strand.}
\keyword{datasets}
