% Generated by roxygen2 (4.0.0): do not edit by hand
\name{viewKegg}
\alias{viewKegg}
\title{Compare KEGG pathways between two functional annotation charts}
\usage{
viewKegg(setA, setB, keggTerm = NULL, species = NULL, workingDir = NULL,
  sortByCount = FALSE, ...)
}
\arguments{
\item{setA}{FunctionalAnnotationChart to compare}

\item{setB}{FunctionalAnnotationChart to compare}

\item{keggTerm}{If a specific KEGG pathway is of interest, input the name here; otherwise,
the most differentially expressed pathway will be used.}

\item{species}{The program can usually figure out the species from the KEGG terms, but if
it can't, supply the species ID here. From pathview vignette,
run 'data(bods); bods' to find species codes.}

\item{workingDir}{The directory to output into. Recommended, since pathview
will put a few different files there each time.}

\item{sortByCount}{Set TRUE if you want the function to automatically choose the pathway
with the most number of genes}

\item{...}{Arguments to be passed to pathview}
}
\value{
Output from pathview: a list of 2, plot.data.gene and plot.data.cpd
}
\description{
viewKegg uses pathview to compare the gene lists visually by KEGG pathway.
You can either supply a pathway id or the function will pick the most differentially
enriched pathway between the two inputs. As functional annotation charts don't have
differential gene expression information, a boolean scale is used - genes in the pathway
are coloured green if from setA, yellow if from both, and red if from setB. We recommend
you supply a working directory, as pathview will download an XML and PNG file as well
as output an additional PNG of the pathway.
}
\examples{
\dontrun{
# Since this function requires writing to a directory, it won't be run here
data(funChart1)
data(funChart2)
viewKegg(funChart1, funChart2)
}
}

