% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Plot dendrogram given an input list of fnAnot charts}
\usage{
plotDendrogram(input)
}
\arguments{
\item{input}{A list of functional annotation charts.}
}
\description{
Given a list of functional annotation charts, 
this function outputs a dendrogram
}

