% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotZScores}
\alias{plotZScores}
\title{Performs z transform on two sets of GO terms and plots scatterplot of result}
\usage{
plotZScores(setA, setB, cutoff = NULL, plotAbs = TRUE, plotNA = FALSE, model = "lm")
}
\arguments{
\item{setA}{DAVIDFunctionalAnnotationChart object to compare}

\item{setB}{DAVIDFunctionalAnnotationChart object to compare}

\item{plotAbs}{Whether to plot the absolute values of z-scores or the raw values}

\item{plotNA}{Whether to remove NAs entirely or set all NAs to 0}

\item{model}{The model to use when plotting linear fit, default 'lm'}

\item{cutoff}{If you want to apply a Benjamini corrected P-value cutoff to each list before generating Z scores, supply it here}
}
\description{
Generates a scatterplot of z transformed GO terms and plots the result along with the Jaccard metric for each GO term and linear fit + correlation.
}
\examples{
data(funChart1)
data(funChart2)
plotZScores(funChart1, funChart2)
}

